+++ server/page/BasePage.php
@@ -367,7 +367,7 @@ abstract class BasePage
      */
     private function getCspRules()
     {
-        $csp_rules = "default-src 'self';  style-src 'self' 'unsafe-inline'; object-src 'none'; script-src 'self' 'unsafe-inline' 'sha256-"
+        $csp_rules = "default-src 'self'; font-src 'self' data:;  style-src 'self' 'unsafe-inline'; object-src 'none'; script-src 'self' 'unsafe-inline' 'sha256-"
             . base64_encode(hash('sha256', $this->get_js_constants(), true)) . "'; img-src 'self' blob: data: https://via.placeholder.com/;";
         return $csp_rules;
     }